Extract the sshTunnelling and execution code to a seperate package and reuse
